second instance

+++ ScreenCapture/Win/WinBase.cpp
@@ -214,7 +214,7 @@ LRESULT WinBase::winProc(HWND hwnd, UINT msg, WPARAM wParam, LPARAM lParam)
         self->onKeyDown(wParam);
     }
     else if (msg == WM_CHAR) {
-        self->onChar(static_cast<UINT>(wParam));
+        self->onChar(wParam);
     }
     else if (msg == WM_IME_STARTCOMPOSITION) {
         self->onIME();
